+++ lizard_wms/models.py
@@ -6,19 +6,29 @@ from urllib import urlencode
 import cgi
 import json
 import logging
+import tls
 import socket
 
+from lizard_wms.conf import settings
 from django.core.urlresolvers import reverse
 from django.db import models
 from django.db import transaction
 from django.template.defaultfilters import urlizetrunc
+from django.utils.html import escapejs
 from django.utils.translation import ugettext_lazy as _
+
 from jsonfield.fields import JSONField
+
 from lizard_map import coordinates
 from lizard_map.models import ADAPTER_CLASS_WMS
+from lizard_map.views import get_view_state
 from lizard_maptree.models import Category
+
 import owslib.wms
 import requests
+
+from lizard_security.manager import FilteredManager
+from lizard_security.models import DataSet
 
 from lizard_wms.widgets import WmsWorkspaceAcceptable
 from lizard_wms import popup_renderers
@@ -163,6 +173,10 @@ class WMSSource(models.Model):
     Definition of a wms source.
     """
 
+    supports_object_permissions = True
+    data_set = models.ForeignKey(DataSet, null=True, blank=True)
+    objects = FilteredManager()
+
     _params = JSONField(
         null=True, blank=True,
         default=WMS_PARAMS_DEFAULT)
@@ -234,12 +248,35 @@ like {"key": "value", "key2": "value2"}.
             # Grmbl, this won't be good for performance.
             return
 
+    def _proxify(self, url):
+        if url is None:
+            return None
+
+        proxied_wms_servers = settings.WMS_PROXIED_WMS_SERVERS
+        for proxied_domain in proxied_wms_servers:
+            if proxied_domain in url:
+                return url.replace(
+                    proxied_domain,
+                    reverse('lizard_wms.wms_proxy', kwargs={
+                            'wms_source_id': self.id}))
+        return url
+
+    @property
+    def proxied_url(self):
+        return self._proxify(self.url)
+
+    @property
+    def proxied_legend_url(self):
+        return self._proxify(self.legend_url)
+
     @property
     def params(self):
         params = {}
         if self._params is not None:
             params = self._params.copy()
         params['layers'] = self.layer_name
+        if 'cql_filter' in params:
+            params['cql_filter'] = escapejs(params['cql_filter'])
         return json.dumps(params)
 
     def update_bounding_box(self, force=False):
@@ -247,7 +284,9 @@ like {"key": "value", "key2": "value2"}.
             try:
                 orig_timeout = socket.getdefaulttimeout()
                 socket.setdefaulttimeout(WMS_TIMEOUT)
-                wms = owslib.wms.WebMapService(self.url, version=FIXED_WMS_API_VERSION)
+                wms = owslib.wms.WebMapService(
+                    self.url,
+                    version=FIXED_WMS_API_VERSION)
                 socket.setdefaulttimeout(orig_timeout)
 
                 params = json.loads(self.params)
@@ -285,9 +324,9 @@ like {"key": "value", "key2": "value2"}.
             adapter_layer_json=json.dumps(
                 {'wms_source_id': self.id,
                  'name': self.layer_name,
-                 'url': self.url,
+                 'url': self.proxied_url,
                  'params': self.params,
-                 'legend_url': self.legend_url,
+                 'legend_url': self.proxied_legend_url,
                  'options': self.options,
                  'cql_filters': list(allowed_cql_filters),
                  'timepositions': self.timepositions,
@@ -298,68 +337,121 @@ like {"key": "value", "key2": "value2"}.
     def capabilities_url(self):
         return capabilities_url(self.url)
 
-    def _bbox_for_feature_info(self, x=None, y=None, radius=None):
-        """Return bbox at point (x,y) in Google coordinates.
-
-        If x, y aren't given, use this layer's bbox, if any. Useful to
-        get available features immediately after fetching the layer.
-        """
-        if x is not None:
-            # Construct the "bounding box", a tiny area around (x,y) We use a
-            # tiny custom radius, because otherwise we don't have enough
-            # control over which feature is returned, there is no mechanism to
-            # choose the feature closest to x, y.
-            if radius is not None:
-                # Adjust the estimated "radius" of an icon on the map.
-                radius /= 50
-                # Convert to wgs84, which is the only supported format for
-                # pyproj.geodesic
-                lon, lat = coordinates.google_to_wgs84(x, y)
-                # Translate center coordinates to lower left and upper right.
-                # Only supports wgs84.
-                # Note: 180 + 45 = 225 = bbox lower left.
-                geod_bbox = coordinates.translate_coords(
-                    [lon] * 2, [lat] * 2, [225, 45], [radius] * 2)
-                # Convert back to web mercator.
-                ll = coordinates.wgs84_to_google(geod_bbox[0][0],
-                                                 geod_bbox[1][0])
-                ur = coordinates.wgs84_to_google(geod_bbox[0][1],
-                                                 geod_bbox[1][1])
-                # Format should be: minX, minY, maxX, maxY.
-                bbox = '{0},{1},{2},{3}'.format(ll[0], ll[1], ur[0], ur[1])
-            else:
-                # Use the old method.
-                fixed_radius = 10
-                bbox = '{0},{1},{2},{3}'.format(
-                    x - fixed_radius, y - fixed_radius,
-                    x + fixed_radius, y + fixed_radius)
-        else:
-            bbox = self.bbox
-        return bbox
-
-    def search_one_item(self, x=None, y=None, radius=None):
+    def search_one_item(self, x=None, y=None, bbox=None,
+                        width=None, height=None, cql_filters=None,
+                        cql_filter_string=None):
         """Return getfeatureinfo values found for a single item."""
-        values = {}
-        bbox = self._bbox_for_feature_info(x=x, y=y, radius=radius)
-        results = self.get_feature_info(bbox=bbox, buffer=16)
+
+        if bbox is None and self.bbox:
+            bbox = self.bbox
+
+        results = self.get_feature_info(
+            bbox=bbox, x=x, y=y, _buffer=16, width=width, height=height,
+            cql_filters=cql_filters,
+            cql_filter_string=cql_filter_string)
+
         # ^^^ Note Reinout: I wonder about that buffer.
+        values = {}
         if results:
             for result in results:
                 values.update(result)
         self._store_features(values)
         return values
 
-    def get_feature_info(self, bbox=None, feature_count=1,
-                         buffer=1, cql_filter_string=None):
+    def _build_payload(self, params, layer, feature_count, version,
+                       bbox, width, height, x, y,
+                       cql_filters, cql_filter_string,
+                       _buffer):
+        """Build the request payload for GetFeatureInfo."""
+
+        payload = {
+            'REQUEST': 'GetFeatureInfo',
+            'EXCEPTIONS': 'application/json',
+            'INFO_FORMAT': 'application/json',
+            'SERVICE': 'WMS',
+            'SRS': 'EPSG:3857',  # Always Google (web mercator)
+            'FEATURE_COUNT': feature_count,
+            # Set the layer we want
+            'LAYERS': layer,
+            'QUERY_LAYERS': layer,
+            'BBOX': bbox,
+            # Height and width in pixels
+            'HEIGHT': height,
+            'WIDTH': width,
+            # The clicked on pixel
+            'X': x,
+            'Y': y,
+
+            # Version from parameter
+            'VERSION': version,
+
+            # Non-standard WMS parameter to slightly increase search
+            # radius.  Shouldn't hurt as most WMS server software ignore
+            # unknown parameters.  see
+            # http://docs.geoserver.org/latest/en/user/services/wms/vendor.html
+            'BUFFER': _buffer,
+            }
+
+        # Add styles to request when defined
+        if 'styles' in params and params['styles']:
+            payload['STYLES'] = params['styles']
+
+        total_cql_filter = []
+        # cql filter string comes from Filter Page
+        if cql_filter_string:
+            total_cql_filter.append(cql_filter_string)
+
+        # cql filter defined in the wms source parameters
+        if 'cql_filter' in params and params['cql_filter']:
+            total_cql_filter.append(params['cql_filter'])
+
+        # CQL filters passed through from the frontend
+        if cql_filters is not None:
+            allowed_filters = self.featureline_set.filter(
+                visible=True, name__in=cql_filters.keys()
+                ).values_list('name', flat=True)
+            for key in allowed_filters:
+                total_cql_filter.append('='.join([key, str(cql_filters[key])]))
+
+        if total_cql_filter:
+            payload['CQL_FILTER'] = ' AND '.join(total_cql_filter)
+
+        # Time selection is added when 'tijd' or 'time' is in the display name
+        if self.timepositions is not None:
+            # Get the user selected date/time selection.
+            date = get_view_state(tls.request)
+            formatting = '%Y-%m-%dT%H:%M:%SZ'
+            payload['TIME'] = '/'.join(
+                d.strftime(formatting)
+                for d in [date['dt_start'], date['dt_end']])
+
+        return payload
+
+    def _parse_response(self, response):
+        if response.status_code != 200:
+            return []
+
+        response_dict = json.loads(response.text)
+        if "exceptions" in response_dict:
+            logger.warning("Error in GetFeatureInfo for layer %s. %s"
+                           % (self.layer_name,
+                              response_dict['exceptions'][0]['text']))
+            return []
+
+        features = response_dict['features']
+        return [obj["properties"] for obj in features]
+
+    def get_feature_info(self, bbox=None, width=1, height=1, x=0, y=0,
+                         feature_count=1, _buffer=1,
+                         cql_filters=None, cql_filter_string=None):
         """Gets feature info from the server inside the bbox.
-
-        Normally the bbox is constructed with ``.bbox_for_feature_info()``.
         """
-
         if not bbox:
             return
-        logger.debug("Getting feature info for %s item(s) in bbox %s",
-                     feature_count, bbox)
+
+        logger.warning("Getting feature info for %s item(s) in bbox %s",
+                       feature_count, bbox)
+
         version = '1.1.1'
         if self.connection and self.connection.version:
             version = self.connection.version
@@ -367,73 +459,16 @@ like {"key": "value", "key2": "value2"}.
         params = json.loads(self.params)
         result = []
         for layer in params['layers'].split(","):
-            payload = {
-                'REQUEST': 'GetFeatureInfo',
-                'EXCEPTIONS': 'application/vnd.ogc.se_xml',
-                'INFO_FORMAT': 'text/plain',
-                'SERVICE': 'WMS',
-                'SRS': 'EPSG:3857',  # Always Google (web mercator)
-                'FEATURE_COUNT': feature_count,
-                # Set the layer we want
-                'LAYERS': layer,
-                'QUERY_LAYERS': layer,
+            payload = self._build_payload(params, layer, feature_count,
+                                          version, bbox, width, height, x, y,
+                                          cql_filters, cql_filter_string,
+                                          _buffer)
+            response = requests.get(self.url, params=payload, timeout=10)
+            layer_result = self._parse_response(response)
 
-                'BBOX': bbox,
-
-                # Get the value at the single pixel of a 1x1 picture
-                'HEIGHT': 1,
-                'WIDTH': 1,
-                'X': 0,
-                'Y': 0,
-
-                # Version from parameter
-                'VERSION': version,
-
-                # Non-standard WMS parameter to slightly increase search
-                # radius.  Shouldn't hurt as most WMS server software ignore
-                # unknown parameters.  see
-                # http://docs.geoserver.org/latest/en/user/services/wms/vendor.html
-                'BUFFER': buffer,
-                # ^^^ Note Reinout: it seems to *greatly* increase search
-                # radius.
-            }
-
-            # Add styles to request when defined
-            if 'styles' in params and params['styles']:
-                payload['STYLES'] = params['styles']
-            if cql_filter_string:
-                payload['CQL_FILTER'] = cql_filter_string
-
-            r = requests.get(self.url, params=payload, timeout=10)
-
-            # XXX Check result code etc
-            if 'no features were found' in r.text:
-                continue
-
-            if not r.text.startswith("Results for FeatureType"):
-                continue
-            # "Parse"
-            one_result = {}
-            for line in r.text.split("\n"):
-                if '----------' in line:
-                    # Store the result, start a new one.
-                    if one_result:
-                        result.append(one_result)
-                    one_result = {}
-                    continue
-                parts = line.split(" = ")
-                if len(parts) != 2:
-                    continue
-                feature, value = parts
-
-                if value.startswith("[GEOMETRY"):
-                    # I think these are always uninteresting
-                    continue
-
-                one_result[feature] = value
             # Store the last result, too, if applicable.
-            if one_result:
-                result.append(one_result)
+            if layer_result:
+                result.extend(layer_result)
         logger.debug("Found %s GetFeatureInfo results.", len(result))
         return result