+++ lizard_wms/views.py
@@ -7,8 +7,11 @@ import logging
 from collections import defaultdict
 
 # from django.utils.translation import ugettext as _
+from lizard_wms.conf import settings
+from django.core.exceptions import PermissionDenied
 from django.core.urlresolvers import reverse
 from django.http import HttpResponse
+from django.http import HttpResponseRedirect
 from django.shortcuts import get_object_or_404
 from django.utils.html import escapejs
 from lizard_map.views import MapView
@@ -16,6 +19,7 @@ from lizard_ui.layout import Action
 import unicodecsv
 
 from lizard_wms import models
+from lizard_wms import url_utils
 
 EMPTY_OPTION = ''
 
@@ -203,3 +207,34 @@ class FilterPageDownload(FilterPageView):
             writer.writerow(feature)
 
         return response
+
+
+def wms_proxy_view(request, wms_source_id):
+    if request.method != 'GET':
+        raise PermissionDenied()
+
+    # If lizard_security decides the current user is not allowed to see
+    # this wms_source, Http404 will be raised
+    wms_source = get_object_or_404(
+        models.WMSSource,
+        pk=wms_source_id)
+
+    url = url_utils.combine_url_and_params(
+        wms_source.url, request.GET)
+
+    # In debug mode, just redirect
+    if settings.DEBUG:
+        return HttpResponseRedirect(url)
+
+    # use Nginx X-Accel-Redirect in production
+    proxied_wms_servers = settings.WMS_PROXIED_WMS_SERVERS
+    for proxied_domain, internal_url in proxied_wms_servers.items():
+        if proxied_domain in url:
+            url = url.replace(
+                proxied_domain,
+                internal_url)
+            break
+
+    response = HttpResponse()
+    response['X-Accel-Redirect'] = url
+    return response